sdks/python/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hyperspacedb"
version = "0.1.0"
description = "Fastest Hyperbolic Vector DB Client"
authors = [
    {name = "YARlabs"}
]
dependencies = [
    "grpcio>=1.50.0",
    "protobuf>=4.21.0",
    "numpy>=1.20.0"
]
requires-python = ">=3.8"

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
cohere = ["cohere>=4.0.0"]
voyage = ["voyageai>=0.1.0"]
google = ["google-generativeai>=0.3.0"]
sentence-transformers = ["sentence-transformers>=2.2.0"]
all = ["openai>=1.0.0", "cohere>=4.0.0", "voyageai>=0.1.0", "google-generativeai>=0.3.0", "sentence-transformers>=2.2.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["hyperspace*"]
